Made terminal more modular, added stack protector, serial i/o support, solved a fuckton of potential bugs

+++ libc/string.h
@@ -38,7 +38,7 @@ extern "C" {
 int memcmp(const void*, const void*, size_t);
 void* memcpy(void* __restrict, const void* __restrict, size_t);
 void* memmove(void*, const void*, size_t);
-void* memset(void*, int, size_t);
+void* memset(void*, uint8_t, size_t);
 void* memsetw(void*, uint16_t, size_t);
 size_t strlen(const char*);
 char *strcpy(char * __restrict dest, const char * __restrict src);
